backend/requirements.txt
flask==2.0.1
flask-restful==0.3.9
google-cloud-firestore==2.1.1
gunicorn==20.1.0
marshmallow==3.12.1
pyjwt==2.1.0
pybase62==0.4.3
requests==2.25.1
spotipy==2.18.0
lyricsgenius==3.0.1
newsapi-python==0.2.6
sparqlwrapper==1.8.5